[CHORE] 소속팀 조회 로직 변경
