+++ src/main/java/com/tiki/server/team/controller/docs/TeamControllerDocs.java
@@ -70,5 +70,34 @@ public interface TeamControllerDocs {
                             description = "서버 내부 오류",
                             content = @Content(schema = @Schema(implementation = ErrorResponse.class)))}
     )
-    public ResponseEntity<SuccessResponse<TeamsGetResponse>> showAllTeam(Principal principal);
+    ResponseEntity<SuccessResponse<TeamsGetResponse>> showAllTeam(Principal principal);
+
+    @Operation(
+            summary = "타임 블록 생성",
+            description = "타임 블록을 생성한다.",
+            responses = {
+                    @ApiResponse(
+                            responseCode = "201",
+                            description = "성공",
+                            content = @Content(schema = @Schema(implementation = SuccessResponse.class))),
+                    @ApiResponse(
+                            responseCode = "403",
+                            description = "접근 권한 없음",
+                            content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+                    @ApiResponse(
+                            responseCode = "404",
+                            description = "유효하지 않은 회원",
+                            content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+                    @ApiResponse(
+                            responseCode = "4xx",
+                            description = "클라이언트(요청) 오류",
+                            content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+                    @ApiResponse(
+                            responseCode = "500",
+                            description = "서버 내부 오류",
+                            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))}
+    )
+    ResponseEntity<SuccessResponse<List<BelongTeamsResponse>>> showBelongTeam(
+            @Parameter(hidden = true) Principal principal
+    );
 }
